+++ coin/src/components/units/main/widget/AvailableWidgets.tsx
@@ -13,5 +13,4 @@ export const availableWidgets = [
     { type: 'longshortratio', name: '바이낸스 선물 비트코인 롱/숏 포지션 비율', icon: <></>, symbol: undefined, category: 'data' },
     { type: 'feargreedindex', name: '공포 / 탐욕 지수', icon: <></>, symbol: undefined, category: 'data' },
     { type: 'btcrsiindex', name: 'BTC RSI', icon: <></>, symbol: undefined, category: 'data' },
-    { type: 'btcstochastic', name: 'BTC 스토캐스틱', icon: <></>, symbol: undefined, category: 'data' },
 ];
